Downloads/fxrates/fxrates/src/test/java/com/viseo/fx/service/UserServiceTest.java
package com.viseo.fx.service;

import com.viseo.fx.config.WebConfig;
import com.viseo.fx.model.User;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.web.WebAppConfiguration;

import static org.junit.Assert.assertTrue;

@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(classes = {WebConfig.class})
@WebAppConfiguration
public class UserServiceTest {

    @Autowired
    private UserService userService;

    @Before
    public void init() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void testGetLatestRates() throws Exception {
        User user = userService.getUserById("101");
        assertTrue("Output message should be match", user.getUserId().equals("101"));
    }
}
